Added init options for contrastive embeddings

+++ src/run_speaking_teemi.sh
@@ -8,7 +8,6 @@ score_names="content pronunciation vocabulary"
 anno_fn="annotation_multi_en_mct_cnn_tdnnf_tgt3meg-dl.xlsx"
 kfold=5
 test_on_valid="true"
-merge_below_b1="false"
 trans_type="trans_stt"
 do_round="true"
 # model-related
@@ -20,6 +19,7 @@ max_seq_length=128
 max_epochs=-1
 alpha=0.5
 num_prototypes=3
+init_prototypes="pretrained"
 monitor="val_score"
 monitor_mode="max"
 model_type=contrastive
@@ -30,6 +30,7 @@ batch_size=8
 accumulate_grad_batches=4
 use_prediction_head=false
 use_pretokenizer=false
+use_layernorm=false
 loss_type="cross_entropy"
 test_book=1
 part=1 # 1 = 基礎聽答, 2 = 情境式提問與問答, 3 = 主題式口說任務, 4 = 摘要報告 (不自動評分) 
@@ -40,7 +41,7 @@ all_bins="1.5,2.5,3.5,4.5,5.5,6.5,7.5"
 cefr_bins="1.5,3.5,5.5,7.5"
 dropout_rate=0.1
 
-extra_options=
+extra_options=""
 
 . ./path.sh
 . ./parse_options.sh
@@ -77,7 +78,12 @@ fi
 if [ "$model_type" == "classification" ] || [ "$model_type" == "regression" ]; then
     exp_tag=${exp_tag}level_estimator_${model_type}
 else
-    exp_tag=${exp_tag}level_estimator_${model_type}_num_prototypes${num_prototypes}
+    if [ $init_prototypes == "pretrained" ]; then
+        exp_tag=${exp_tag}level_estimator_${model_type}_num_prototypes${num_prototypes}
+    else
+        extra_options="$extra_options --init_prototypes ${init_prototypes}"
+        exp_tag=${exp_tag}level_estimator_${model_type}_num_prototypes${num_prototypes}_${init_prototypes}
+    fi
 fi
 
 if [ "$do_loss_weight" == "true" ]; then
@@ -100,6 +106,11 @@ if [ "$use_pretokenizer" == "true" ]; then
     extra_options="$extra_options --use_pretokenizer"
 fi
 
+if [ "$use_layernorm" == "true" ]; then
+    exp_tag=${exp_tag}_lnorm
+    extra_options="$extra_options --use_layernorm"
+fi
+
 if [ "$loss_type" != "cross_entropy" ]; then
     exp_tag=${exp_tag}_${loss_type}
     extra_options="$extra_options --loss_type $loss_type"
@@ -108,6 +119,7 @@ fi
 if [ "$max_epochs" != "-1" ]; then
     exp_tag=${exp_tag}_ep${max_epochs}
 fi
+
 if [ "$max_epochs" != "-1" ]; then
     exp_tag=${exp_tag}_ep${max_epochs}
 fi
@@ -115,7 +127,7 @@ fi
 model_name=`echo $model_path | sed -e 's/\//-/g'`
 exp_tag=${exp_tag}_${model_name}_${monitor}-${monitor_mode}_b${batch_size}g${accumulate_grad_batches}_lr${init_lr}_drop${dropout_rate}
 
-if [ $stage -le 0 ] && [ $stop_stage -ge 0 ]; then       
+if [ $stage -le 0 ] && [ $stop_stage -ge 0 ]; then  
     for sn in $score_names; do
         for fd in $folds; do
             echo "$part $sn $fd $exp_tag"
@@ -136,9 +148,9 @@ if [ $stage -le 0 ] && [ $stop_stage -ge 0 ]; then
                                       --out $exp_root \
                                       --exp_dir $exp_dir \
                                       --score_name $sn \
-                                      --dropout_rate ${dropout_rate} \
                                       --batch $batch_size --warmup 0 \
                                       --accumulate_grad_batches $accumulate_grad_batches \
+                                      --dropout_rate $dropout_rate \
                                       --num_prototypes $num_prototypes --type ${model_type} --init_lr $init_lr \
                                       --alpha $alpha --data $data_dir/$fd --test $data_dir/$fd 
         done
@@ -153,6 +165,11 @@ if [ $stage -le 1 ] && [ $stop_stage -ge 1 ]; then
             # Test a pretrained model
             checkpoint_path=`find $exp_root/$exp_tag/$sn/$fd/version_0 -name *ckpt`
             
+            if [ -z $checkpoint_path ]; then
+                echo "No such directories, $exp_root/$exp_tag/$sn/$fd/version_0";
+                exit 0;
+            fi
+            
             if [ -d $exp_root/$exp_tag/$sn/$fd/version_1 ]; then
                 rm -rf $exp_root/$exp_tag/$sn/$fd/version_1
             fi
@@ -160,7 +177,7 @@ if [ $stage -le 1 ] && [ $stop_stage -ge 1 ]; then
             echo "$part $sn $fd"
             echo $checkpoint_path
             exp_dir=$exp_tag/$sn/$fd
-            python level_estimator.py --model $model_path --lm_layer 11 $extra_options \
+            python level_estimator.py --model $model_path --lm_layer 11 $extra_options --do_test \
                                       --CEFR_lvs  $max_score \
                                       --seed 985 --num_labels $max_score \
                                       --max_epochs $max_epochs \
@@ -169,7 +186,6 @@ if [ $stage -le 1 ] && [ $stop_stage -ge 1 ]; then
                                       --exp_dir $exp_dir \
                                       --score_name $sn \
                                       --batch $batch_size --warmup 0 \
-                                      --accumulate_grad_batches $accumulate_grad_batches \
                                       --num_prototypes $num_prototypes --type ${model_type} --init_lr $init_lr \
                                       --alpha $alpha --data $data_dir/$fd --test $data_dir/$fd --out $exp_root --pretrained $checkpoint_path
 
@@ -184,7 +200,7 @@ if [ $stage -le 3 ] && [ $stop_stage -ge 3 ]; then
                                                     --all_bins "$all_bins" \
                                                     --cefr_bins "$cefr_bins" \
                                                     --folds "$folds" \
-                                                    --version_dir version_0 \
+                                                    --version_dir version_1 \
                                                     --scores "$score_names" > $runs_root/$exp_tag/report.log
     
 fi
@@ -206,6 +222,7 @@ if [ $stage -le 5 ] && [ $stop_stage -ge 5 ]; then
                                                     --cefr_bins "$cefr_bins" \
                                                     --folds "$folds" \
                                                     --question_type tb${test_book}p${part} \
+                                                    --version_dir version_1 \
                                                     --scores "$score_names" > $runs_root/$exp_tag/report_spk.log
 fi
